Show notification on startup whether update succeeded or not. Cancel update when internet connection is too bad, i.e. update takes longer than 7 seconds. #10

+++ DdfGuide.Android/ToastNotifier.cs
@@ -15,7 +15,7 @@ namespace DdfGuide.Android
 
         public void Notify(string message)
         {
-            var toast = Toast.MakeText(_context, message, ToastLength.Short);
+            var toast = Toast.MakeText(_context, message, ToastLength.Long);
             toast.Show();
         }
     }
